Issue-#124: Covenant commands split into separate files
A lot of changes, but mostly separating covenant command into separate classes and files.
