+++ src/ChatBot.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 4.0.3, 2017-05-24T00:07:33. -->
+<!-- Written by QtCreator 4.0.3, 2017-05-28T03:00:01. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
