+++ src/Utils/Config/ConfigurationParameters.hpp
@@ -27,6 +27,7 @@
 /* Covenant params */
 #define CFGP_COV_JOIN_PRICE         "CovJoinPrice"
 #define CFGP_COV_CREATE_PRICE       "CovCreatePrice"
+#define CFGP_COV_RENAME_PRICE       "CovRenamePrice"
 #define CFGP_VGRAPH_UPD_TIME        "ViewerGraphUpdateTime"
 /* Message params */
 #define CFGP_MESSAGE_GRAPH_UPD_TIME "MessagegraphUpdateTime"
